Several package moves, as part of the `org.bdgenomics.adam.util` cleanup:
* Moved `NormalizationUtils` to `org.bdgenomics.adam.algorithms.consensus`, and
  made it package private to `consensus`.
* Moved `MdTag` to `org.bdgenomics.adam.models`.
* Moved `QualityScore` to `org.bdgenomics.adam.models` and made it package
  private to `consensus`.
