+++ adam-core/src/main/scala/org/bdgenomics/adam/rdd/read/recalibration/Covariate.scala
@@ -17,8 +17,8 @@
  */
 package org.bdgenomics.adam.rdd.read.recalibration
 
+import org.bdgenomics.adam.models.QualityScore
 import org.bdgenomics.adam.rich.DecadentRead
-import org.bdgenomics.adam.util.QualityScore
 
 /**
  * A Covariate represents a predictor, also known as a "feature" or
