+++ adam-core/src/main/scala/org/bdgenomics/adam/rdd/read/recalibration/Recalibrator.scala
@@ -17,10 +17,10 @@
  */
 package org.bdgenomics.adam.rdd.read.recalibration
 
+import org.bdgenomics.adam.models.QualityScore
 import org.bdgenomics.adam.rich.DecadentRead.Residue
 import org.bdgenomics.adam.rich.RichAlignmentRecord._
 import org.bdgenomics.adam.rich.DecadentRead
-import org.bdgenomics.adam.util.QualityScore
 import org.bdgenomics.formats.avro.AlignmentRecord
 import org.bdgenomics.adam.instrumentation.Timers._
 import scala.math.{ exp, log }
