+++ adam-core/src/test/scala/org/bdgenomics/adam/algorithms/consensus/NormalizationUtilsSuite.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.adam.util
+package org.bdgenomics.adam.algorithms.consensus
 
 import htsjdk.samtools.TextCigarCodec
 import org.bdgenomics.adam.rich.RichAlignmentRecord._
